% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/amca.R
\name{amca}
\alias{amca}
\title{AMCA algorithm.}
\usage{
amca(DATA, parameters, MPIs, ResultsFolder, selectedModels = NULL,
  warmup = NULL, verbose = FALSE, PreSel = TRUE, allBounds = FALSE)
}
\arguments{
\item{DATA}{This is a data.frame containing the observed time series (zoo objects). It is structured into three columns containing: precipitation (P), potential evapo-transpiration (E) and streamflow discharge (Q).}

\item{parameters}{This is a named data frame containing the parameter table, where each column corresponds to a parameter and each row to a realization.}

\item{MPIs}{list of functions describing the Model performance Indices.}

\item{ResultsFolder}{Path to the folder containing results from MC simulations.}

\item{selectedModels}{(OPTIONAL) This is a table that contains at list 1 column named "mid" (list of model structures). Other informations can be added as additional columns but will be ignored (default = NULL).}

\item{warmup}{Number of initial time steps to ignore (default = 10 percent of the length of DATA).}

\item{verbose}{if set to TRUE it prints running information (default = FALSE).}

\item{PreSel}{if set to FALSE the preliminary selection step is skipped (default = TRUE).}

\item{allBounds}{if set to TRUE it calculates ensembles of intermediate steps, if set to FALSE only BoundIE and BoundsRE are calculated (default = FALSE). In this case the MC simulations should produce 2 ojects: indices and simulates discharge table.}
}
\value{
A list of objects to infer.
}
\description{
AMCA algorithm.
}
\examples{
# results <- damach( DATA, parameters, MPIs, ResultsFolder )
}

