% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BarPlots.R
\name{BarPlots}
\alias{BarPlots}
\title{Create bar plots}
\usage{
BarPlots(mTable, selectedOption)
}
\arguments{
\item{mTable}{molt table}

\item{selectedOption}{a selection of options}
}
\value{
bar plot
}
\description{
Create bar plots
}
\examples{
# BarPlots(mTable,selectedOption)
}

