% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BuildEnsemble}
\alias{BuildEnsemble}
\title{Extracts the ensemble bounds from a given a set of realisations.}
\usage{
BuildEnsemble(DATA, warmup, Realisations2Use, SimulationFolder,
  maxminOnly = TRUE, lowerP = 0.05, upperP = 0.95, verbose = FALSE)
}
\arguments{
\item{DATA}{this is the data frame containing the observations}

\item{warmup}{this is the warmup period}

\item{Realisations2Use}{list of the filtered realisations}

\item{SimulationFolder}{path to the folder where simulation results are stored}

\item{maxminOnly}{if set to FALSE it calculates max/min bounds and upper/lower percentile bounds, otherwise only the max/min bounds (default is FALSE).}

\item{lowerP}{lower probability (e.g. 0.05 means 5th percentile)}

\item{upperP}{upper probability (e.g. 0.95 means 95th percentile)}

\item{verbose}{if set to TRUE it prints running information}
}
\value{
A data.frame with 6 columns: date&time (Dates), observed discharge (Qobs), lower bound (LB), upper bound (UB), lower percentile (LP), upper percentile (UP)
}
\description{
Results are generally divided in multiple files, therefore the function requires the simulations' folder path.
}
\examples{
# BuildEnsemble(DATA, warmup, Realisations2Use, SimulationFolder)
}

