% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CompareModelOptions.R
\name{CompareModelOptions}
\alias{CompareModelOptions}
\title{Compare AMCA runs}
\usage{
CompareModelOptions(listResults, experimentName = "A")
}
\arguments{
\item{listResults}{list of multiple amca runs}

\item{experimentName}{name of the experiment}
}
\value{
data.frame with summarized info
}
\description{
Compare AMCA runs
}
\examples{
# CompareModelOptions(listResults, experimentName="A")
}

