% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleForecast.R
\name{EnsembleForecast}
\alias{EnsembleForecast}
\title{Plot QQ and results of statistical tests.}
\usage{
EnsembleForecast(simulated, observed, warmupDTW = 0, verbose = FALSE)
}
\arguments{
\item{simulated}{table m(rows) x n(cols) containing m predictions for each n time steps}

\item{observed}{observed discharge time series}

\item{warmupDTW}{additional warmup period (optional)}

\item{verbose}{if set to TRUE it prints running information}
}
\value{
A plot with QQ plots and results of statistical tests
}
\description{
Plot QQ and results of statistical tests.
}
\examples{
# EnsembleForecast(predDTW,observed,warmupDTW=0,verbose=FALSE)

}

