% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendTable.R
\name{ExtendTable}
\alias{ExtendTable}
\title{Extends the table of realisations to include the indices, model building options and parameters.}
\usage{
ExtendTable(realisations, ModelList, Indices, parameters, ObsIndicesNames,
  verbose = FALSE, onlyIndices = FALSE)
}
\arguments{
\item{realisations}{table containing at least 2 columns: mid (list of model structures) and pid (parameter sets)}

\item{ModelList}{a table that contains at list 1 colum named "mid" (list of model structures). Other informations can be added as additional columns but will be ignored.}

\item{Indices}{this is the array containing the Model Performance Indices rescaled in the range [0,1], see \code{RescaleIndices}.}

\item{parameters}{This is a named data frame containing the parameter table, where each column corresponds to a parameter and each row to a realization.}

\item{ObsIndicesNames}{this is the list of model performance indices calculated from observations.}

\item{verbose}{if set to TRUE it prints running information}

\item{onlyIndices}{if set to TRUE to extends the table of realisations only with indices.}
}
\value{
Extended data.frame
}
\description{
Extends the table of realisations to include the indices, model building options and parameters.
}
\examples{
# ExtendTable(PreSel,ModelList,Indices,parameters,ObsIndicesNames)

}
