% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FindBreaksLabels}
\alias{FindBreaksLabels}
\title{Utility function to define breaks for plotting modelling options}
\usage{
FindBreaksLabels(modelOption)
}
\arguments{
\item{modelOption}{modelling option}
}
\value{
breaks
}
\description{
Utility function to define breaks for plotting modelling options
}
\examples{
# FindBreaksLabels(modelOption)
}

