% Generated by roxygen2 (4.0.2): do not edit by hand
\name{KirpichTc}
\alias{KirpichTc}
\title{This function calculates the catchment's time of concentration (Tc) using Kirpich's formula.}
\usage{
KirpichTc(length, slope)
}
\arguments{
\item{length}{this is the length of the main river channel (in metres, usually obtained by GIS layers)}

\item{slope}{this is the average slope of the main river channel (dimensionless, usually obtained by GIS layers)}
}
\value{
time of concentration (in days)
}
\description{
This function calculates the catchment's time of concentration (Tc) using Kirpich's formula.
}
\examples{
# For Severn catchment (mid-Wales, UK)
# Tc <- KirpichTc( length=4601, slope=0.065 )
}

