% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LabelOptions.R
\name{LabelOptions}
\alias{LabelOptions}
\title{Utility function to define labels for plotting modelling options}
\usage{
LabelOptions(numericalOption)
}
\arguments{
\item{numericalOption}{numerical option defining the compone component}
}
\value{
labels
}
\description{
Utility function to define labels for plotting modelling options
}
\examples{
# LabelOptions(numericalOption)
}

