% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MergeRE.R
\name{MergeRE}
\alias{MergeRE}
\title{Merge Reduced Ensembles.}
\usage{
MergeRE(periodRuns, path, parameters)
}
\arguments{
\item{periodRuns}{vector containing the years to merge}

\item{path}{directory path containing AMCA results}

\item{parameters}{data frame containing all the sampled parameters}
}
\value{
A list of objects to infer.
}
\description{
Merge Reduced Ensembles.
}
\examples{
# MergeRE( periodRuns, path )
}

