% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFrontier.R
\name{ParetoFrontier}
\alias{ParetoFrontier}
\title{Generate a table containing only non-dominated realisations based on MPIs and their Pareto Frontier.}
\usage{
ParetoFrontier(realisations, ObsIndicesNames = c("LAGTIME", "MAE", "NSHF",
  "NSLF"))
}
\arguments{
\item{realisations}{list of model structures to take into account}

\item{ObsIndicesNames}{this is a vector containing the names of the indices.}
}
\value{
a subset of the initial dataset containing only non-dominated realisations
}
\description{
Generate a table containing only non-dominated realisations based on MPIs and their Pareto Frontier.
}
\examples{
# ParetoFrontier(PreSelTable)

}

