% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ParetoFrontierDTW}
\alias{ParetoFrontierDTW}
\title{Add columns with cluster information to the Pareto Front}
\usage{
ParetoFrontierDTW(ParetoFront, DATA, the.som, SimulationFolder, observedQ)
}
\arguments{
\item{ParetoFront}{Pareto Frontier}

\item{DATA}{this is the data frame containing the observations}

\item{the.som}{som map}

\item{SimulationFolder}{path to the folder where the function saves the results}

\item{observedQ}{observed discharge (warmup removed)}
}
\value{
A data.frame with 4 columns: date&time (Dates), observed discharge (Qobs), lower bound (LB), upper bound (UB)
}
\description{
Add columns with cluster information to the Pareto Front
}
\details{
The table containing the pareto frontier is now extended with clustering information.
}
\examples{
# ParetoFrontierDTW(ParetoFront,DATA,the.som,SimulationFolder,observedQ)
}

