% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PerformanceIndices}
\alias{PerformanceIndices}
\title{Calculate 5 Model Performance Indices (MPIs)}
\usage{
PerformanceIndices(Po, Qo, Qs)
}
\arguments{
\item{Po}{observed precipitation [mm/day]}

\item{Qo}{observed streamflow discharge [mm/day]}

\item{Qs}{simulated streamflow discharge [mm/day]}
}
\value{
A list of 5 indices: LAGTIME, MAE, NSHF, NSLF, RR.
}
\description{
Calculate 5 Model Performance Indices (MPIs)
}
\details{
LAGTIME is calculated from the tiger::lagtime function. MAE is the mean absolute error. NSHF is the complement to 1 of Nash-Sutcliffe efficiency for high flows. NSLF is the complement to 1 of Nash-Sutcliffe efficiency for low flows. RR is the Rainfall-Runoff coefficient. All the indices above tend to zero (best performance).
}
\examples{
# PerformanceIndices(Po,Qo,Qs)
}

