% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersistenceRate.R
\name{PersistenceRate}
\alias{PersistenceRate}
\title{Calculate persistance rate}
\usage{
PersistenceRate(REtable, ModelList, modellingOptions)
}
\arguments{
\item{REtable}{summary table of reduced ensemble}

\item{ModelList}{this is the list of model structures}

\item{modellingOptions}{vector containing the names of modelling options (default is c("rferr","arch1","arch2","qsurf","qperc","esoil","qintf","q_tdh"))}
}
\value{
data.frame containing the persistence rate in a new column
}
\description{
Calculate persistance rate
}
\examples{
# PersistenceRate(REtable,ModelList,modellingOptions)

}
