% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Plot2DarrayMultiple}
\alias{Plot2DarrayMultiple}
\title{This function plots the 2Darray of a certain index.}
\usage{
Plot2DarrayMultiple(indices, mList)
}
\arguments{
\item{indices}{arrayP containing the performance measures for all the realisations}

\item{mList}{This is a table which contains at least 1 colum named "mid". Other informations can be added as additional columns but will be ignored.}
}
\value{
A grid of matrix plots with color coded cells.
}
\description{
This function plots the 2Darray of a certain index.
}
\examples{
# Plot2DarrayMultiple(Indices,mList)
}

