% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Plot2DarraySingle.R
\name{Plot2DarraySingle}
\alias{Plot2DarraySingle}
\title{This function plots the 2Darray of a certain index.}
\usage{
Plot2DarraySingle(indices, mList, idx)
}
\arguments{
\item{indices}{arrayP containing the performance measures for all the realisations}

\item{mList}{This is a table which contains at least 1 colum named "mid". Other informations can be added as additional columns but will be ignored.}

\item{idx}{is the index number}
}
\value{
A matrix plot with color coded cells.
}
\description{
This function plots the 2Darray of a certain index.
}
\examples{
# Plot2DarraySingle(Indices,mList,idx)
}

