% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotCube.R
\name{PlotCube}
\alias{PlotCube}
\title{Plot cube to visualise 3D array}
\usage{
PlotCube(myDimensions = NULL)
}
\arguments{
\item{myDimensions}{vector listing the number of elements in each dimension: number of parameter stes, number of indices and number of model structures. By default this is c(10000,5,1248).}
}
\value{
A plot showing the 3D array
}
\description{
Plot cube to visualise 3D array
}
\examples{
# PlotCube()
}

