% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotData}
\alias{PlotData}
\title{Plot input data}
\usage{
PlotData(DATA)
}
\arguments{
\item{DATA}{This is a data.frame containing the observed time series (zoo objects).
            It is structured into three columns:
            "P" containing precipitation,
            "E" containing evapo-transpiration and
            "Q" containing streamflow discharge.}
}
\value{
A plot divided into 3 panels: P, E andQ
}
\description{
Plot input data
}
\examples{
# PlotData(DATA)
}

