% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotEnsembles.R
\name{PlotEnsembles}
\alias{PlotEnsembles}
\title{Compare 2 distributions percentiles over time.}
\usage{
PlotEnsembles(bounds, dischargeTable, lowerP = 5, upperP = 95,
  label1 = "T limits", label2 = "T' percentiles", type = "B")
}
\arguments{
\item{bounds}{This is a data.frame structured into five columns:
"Dates" containing timestamps,
"Qobs" containing observed streamflow discharge,
"LB" lower bound (minimum) of the set of realisations,
"UB" upper bound (maximum) of the set of realisations,
"LP" lower percentile of the set of realisations,
"UP" upper percentile of the set of realisations.}

\item{dischargeTable}{This is a numerical matrix structured into R number of rows and C number of columns. R corresponds to the number of realisations and C corresponds to the number of timesteps.}

\item{lowerP}{lower percentile to visualise in red (default = 5)}

\item{upperP}{upper percentile to visualise in red (default = 95)}

\item{label1}{label for the x-axis.}

\item{label2}{label for the y-axis.}

\item{type}{If "B" means the reference is min-max. If "P" it means the reference is lower-upper percentile.}
}
\value{
A plot comparing ensemble distributions.
}
\description{
Compare 2 distributions percentiles over time.
}
\examples{
# PlotModelSimilarities(ModelList,EnsembleTable)
}

