% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotMedians}
\alias{PlotMedians}
\title{Plot the 1 panel for each MPI, showing the median for each MID.}
\usage{
PlotMedians(ModelList, Indices)
}
\arguments{
\item{ModelList}{This is a table which contains at least 1 colum named "mid". Other informations can be added as additional columns but will be ignored.}

\item{Indices}{arrayP containing the performance measures for all the realisations}
}
\value{
A plot showing median values for each MPI and MID.
}
\description{
Plot the 1 panel for each MPI, showing the median for each MID.
}
\examples{
# PlotMedians <- function(ModelList,Indices)
}

