% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotModelSimilarities}
\alias{PlotModelSimilarities}
\title{Plot model similarities (for comparison).}
\usage{
PlotModelSimilarities(ModelList, EnsembleTable, synMID = "",
  plotType = "frequency", myTitle = "")
}
\arguments{
\item{ModelList}{This is a table which contains at least 1 colum named "mid". Other informations can be added as additional columns but will be ignored.}

\item{EnsembleTable}{table containing alle the combination of mid and pid selected so far}

\item{synMID}{if running synthetic experiments, this is the synthetic MID number (otherwise leave it empty)}

\item{plotType}{default is "frequency", TODO: "persistency"}

\item{myTitle}{title of the plot}
}
\value{
A plot comparing model similarities.
}
\description{
Plot model similarities (for comparison).
}
\examples{
# PlotModelSimilarities(ModelList,EnsembleTable)
}

