% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotParameterSimilarities}
\alias{PlotParameterSimilarities}
\title{Plot prior and posterior parameter distributions (for comparison).}
\usage{
PlotParameterSimilarities(RealisationsTable, parameters,
  selectedParams = NULL, synparameters = NULL)
}
\arguments{
\item{RealisationsTable}{list of the filtered realisations}

\item{parameters}{This is a named data frame containing the parameter table, where each column corresponds to a parameter and each row to a realization.}

\item{selectedParams}{is a list of pid to filter}

\item{synparameters}{is the synthetic parameter set (if available)}
}
\value{
A plot comparing prior and posterior parameter distributions.
}
\description{
Plot prior and posterior parameter distributions (for comparison).
}
\examples{
# Plot_ParameterSimilarities(RealisationsTable, parameters)
}

