% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pNS.R
\name{pNS}
\alias{pNS}
\title{Probabilistic Nash-Sutcliffe efficiency}
\usage{
pNS(Qobs, discharges)
}
\arguments{
\item{Qobs}{vector of observed values}

\item{discharges}{is a matrix of simulations}
}
\value{
probabilistic efficiency
}
\description{
This is an implementation of the probabilistic Nash-Sutcliffe efficiency defined by Bulygina et al. (2009).
}
\examples{
# pNS(Qobs, discharges)

}
