% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision}
\alias{precision}
\title{Calculate precision/sharpness}
\usage{
precision(lowerBound1, upperBound1, lowerBound2, upperBound2)
}
\arguments{
\item{lowerBound1}{vector containing the minimum or the lower quantile bound for the first ensemble}

\item{upperBound1}{vector containing the maximum or the upper quantile bound for the first ensemble}

\item{lowerBound2}{vector containing the minimum or the lower quantile bound for the second ensemble}

\item{upperBound2}{vector containing the maximum or the upper quantile bound for the second ensemble}
}
\value{
Returns precision, in the range [0,100], percentage of closeness to a line.
}
\description{
Precision definition available in Yadav et al. (2007)
}
\examples{
# precision(lowerBound1, upperBound1, lowerBound2, upperBound2)

}
