% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PredictionsExtract}
\alias{PredictionsExtract}
\title{Extract the ensemble from a given a set of realisations.}
\usage{
PredictionsExtract(Realisations2Use, SimulationFolder, SimLength, warmup)
}
\arguments{
\item{Realisations2Use}{list of the filtered realisations}

\item{SimulationFolder}{path to the folder where simulation results are stored}

\item{SimLength}{length of the simulation results, usually it is total length - warmup}

\item{warmup}{this is the warmup period}
}
\value{
predictions
}
\description{
Results are generally divided in multiple files, therefore the function requires the simulations' folder path.
}
\examples{
#
}

