% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareTable.R
\name{PrepareTable}
\alias{PrepareTable}
\title{Prepare a table containing Model ID numbers, model options and parameter in use.}
\usage{
PrepareTable(reduced = FALSE)
}
\arguments{
\item{reduced}{boolean value. If FALSE the entire model list is taken into account. If FALSE, a reduced version is selected.}
}
\value{
table containing Model ID numbers, model options and parameter in use
}
\description{
Prepare a table containing Model ID numbers, model options and parameter in use.
}
\examples{
# PrepareTable()

}
