% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PreSelection}
\alias{PreSelection}
\title{Perform the preliminary selection}
\usage{
PreSelection(ModelList, Indices, threshold, selectM = TRUE, selectP = TRUE,
  type = "independent")
}
\arguments{
\item{ModelList}{this is the list of model structures}

\item{Indices}{this is the array containing the Model Performance Indices rescaled in the range [0,1], see \code{RescaleIndices}.}

\item{threshold}{is a number between 0 and 1 setting the filtering threshold.}

\item{selectM}{boolean value. If it is TRUE the preselection applies to model structures}

\item{selectP}{boolean value. If it is TRUE the preselection applies to parameter sets}

\item{type}{it can either be "dependent" or "independent"}
}
\value{
list of two data.frames, each with two columns:
models = row number (row) and model identification number (mid);
params = row number (row) and model identification number (pid), by definition row and pid coincide.
}
\description{
Perform the preliminary selection
}
\examples{
# PreSelection(ModelList, Indices, threshold)
}

