% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedundancyReduction.R
\name{RedundancyReduction}
\alias{RedundancyReduction}
\title{Redundancy reduction process using SOM clustering and Dynamic Time Warping.}
\usage{
RedundancyReduction(PF, observedQ, ResultsFolder, ObsIndicesNames,
  verbose = FALSE)
}
\arguments{
\item{PF}{Pareto Frontier}

\item{observedQ}{observed discharge (warmup removed)}

\item{ResultsFolder}{Path to the folder containing results from MC simulations.}

\item{ObsIndicesNames}{is the list of indices names.}

\item{verbose}{if set to TRUE it prints running information (default = FALSE)}
}
\value{
A subset of the table containing the filtered configurations.
}
\description{
Redundancy reduction process using SOM clustering and Dynamic Time Warping.
}
\details{
This function removes redundant realizations by extracting the realization with the best DTW score from each cluster.
}
\examples{
# RedundancyReduction(PF, observedQ, ResultsFolder, ObsIndicesNames)

}

