% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RedundancyReductionDTW}
\alias{RedundancyReductionDTW}
\title{Redundancy reduction process using Dynamic Time Warping.}
\usage{
RedundancyReductionDTW(DTWtable)
}
\arguments{
\item{DTWtable}{pareto frontier of the DTW realisations}
}
\value{
A subset of the table containing the Pareto frontier.
}
\description{
Redundancy reduction process using Dynamic Time Warping.
}
\details{
This function removes redundant realizations by extracting the realization with the best DTW score from each cluster.
}
\examples{
# ParetoFrontier(indices[,,selected_models[,"row"]],selected_models,parameters,SelectedIndices)
}

