% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RelativeFrequency}
\alias{RelativeFrequency}
\title{Calculate relative frequency}
\usage{
RelativeFrequency(REtable, modellingOptions)
}
\arguments{
\item{REtable}{summary table of reduced ensemble}

\item{modellingOptions}{vector containing the names of modelling options (default is c("rferr","arch1","arch2","qsurf","qperc","esoil","qintf","q_tdh"))}
}
\value{
data.frame containing the relative frequency in a new column
}
\description{
Calculate relative frequency
}
\examples{
# RelativeFrequency(REtable,modellingOptions)
}

