% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RescaleIndices.R
\name{RescaleIndices}
\alias{RescaleIndices}
\title{Rescale Model Performance Indices (MPIs) between 0 and 1.}
\usage{
RescaleIndices(Indices)
}
\arguments{
\item{Indices}{this is the array containing the not-yet rescaled Model Performance Indices.}
}
\value{
Rescaled array of Model Performance Indices
}
\description{
Rescale Model Performance Indices (MPIs) between 0 and 1.
}
\details{
The best performance always tends to zero.
}
\examples{
# newIndices <- RescaleIndices(Indices)

}
