% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{review}
\alias{review}
\title{Summarise performance coefficients}
\usage{
review(IE, RE)
}
\arguments{
\item{IE}{Initial Ensemble bounds}

\item{RE}{Reduced Ensemble bounds}
}
\value{
Returns accuracy, precision and statistical reliability
}
\description{
Summarise performance coefficients
}
\examples{
# review(IE, RE)

}

