% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowMatch.R
\name{RowMatch}
\alias{RowMatch}
\title{Match rows in two matrices}
\usage{
RowMatch(A, B)
}
\arguments{
\item{A}{first matrix}

\item{B}{second matrix}
}
\value{
matching row
}
\description{
Match rows in two matrices
}
\examples{
# ParetoFrontier(indices[,,selected_models[,"row"]],selected_models,parameters,SelectedIndices)

}

