% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ScanResultFiles}
\alias{ScanResultFiles}
\title{This function scans the result folder for missing files.}
\usage{
ScanResultFiles(ResultFolder, ModelList)
}
\arguments{
\item{ResultFolder}{path to the folder containing the results (simulations)}

\item{ModelList}{list of model structure to take into account}
}
\value{
prints on the screen two lists: 1) MID of existing files, 2) MID of missing files
}
\description{
This function scans the result folder for missing files.
}
\examples{
# ScanResultFiles("/test",damachModelList)
}

