% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ScenarioPlotDTW}
\alias{ScenarioPlotDTW}
\title{Extract (from the object discharges) the time series for a certain realization}
\usage{
ScenarioPlotDTW(mid, pid, SimulationFolder)
}
\arguments{
\item{mid}{model structure identification number}

\item{pid}{parameter set identification number}

\item{SimulationFolder}{path to the folder where the function saves the results}
}
\value{
subset of discharges table
}
\description{
Extract (from the object discharges) the time series for a certain realization
}
\examples{
# ScenarioPlotDTW(mid,pid,SimulationFolder)
}

