% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectIndices.R
\name{SelectIndices}
\alias{SelectIndices}
\title{Finds Indices for selected realisations.}
\usage{
SelectIndices(realisations, Indices, ModelList, ObsIndicesNames)
}
\arguments{
\item{realisations}{table containing at least 2 columns: mid (list of model structures) and pid (parameter sets)}

\item{Indices}{this is the array containing the rescaled Model Performance.}

\item{ModelList}{this is the list of model structures.}

\item{ObsIndicesNames}{this is the list of model performance indices calculated from observations.}
}
\value{
Extended data.frame
}
\description{
Finds Indices for selected realisations.
}
\examples{
# SelectIndices(realisations)

}
