% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Simulations2Indices}
\alias{Simulations2Indices}
\title{Collate indices of various model structures in one 3-dimensional array.}
\usage{
Simulations2Indices(ModelList, ResultsFolder, nParams, nIndices, verbose)
}
\arguments{
\item{ModelList}{this is the reduced list of model structures (total number: 252)}

\item{ResultsFolder}{path to the folder where the function saves the results}

\item{nParams}{number of sampled parameter sets}

\item{nIndices}{number of Model Performance Indices (MPIs)}

\item{verbose}{if TRUE, it prints details of the current simulation}
}
\value{
3D arrays containing in the first dimension the parameters, in the second dimension the indices and in the third dimension the model structures.
}
\description{
Collate indices of various model structures in one 3-dimensional array.
}
\examples{
# indices <- Simulations2Indices(ModelList,ResultsFolder,nParams, nIndices, verbose)
}

