% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SplitYears.R
\name{SplitSolarYears}
\alias{SplitSolarYears}
\title{Split dataset into solar years (from 1st January to the following 31st December)}
\usage{
SplitSolarYears(myDATA, myYear = NULL)
}
\arguments{
\item{myDATA}{it is a dataframe containing three zoo time series: P (precipitation), E(evapotranspiration) and Q(discharge)}

\item{myYear}{number identifying a particular year to output. If this is not null, the year is returned and not saved in a file.}
}
\value{
saves in the working directory an rda file for each year
}
\description{
Split dataset into solar years (from 1st January to the following 31st December)
}

