% Generated by roxygen2 (4.0.2): do not edit by hand
\name{amca}
\alias{amca}
\title{Run the AMCA algorithm.}
\usage{
amca(DATA, ModelList, warmup, parameters, ResultsFolder, verbose = FALSE)
}
\arguments{
\item{DATA}{This is a data.frame containing the observed time series (zoo objects).
It is structured into three columns:
"P" containing precipitation,
"E" containing evapo-transpiration and
"Q" containing streamflow discharge.}

\item{ModelList}{a table that contains at list 1 colum named "mid" (list of model structures). Other informations can be added as additional columns but will be ignored.}

\item{warmup}{Number of initial time steps to ignore.}

\item{parameters}{This is a named data frame containing the parameter table, where each column corresponds to a parameter and each row to a realization.}

\item{ResultsFolder}{Path to the folder containing results from MC simulations.}

\item{verbose}{if set to TRUE it prints running information}
}
\value{
A list of objects to infer.
}
\description{
Run the AMCA algorithm.
}
\examples{
# results <- damach(DATA, ModelList, warmup, parameters, ResultsFolder)
}

