% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amca.R
\name{amca}
\alias{amca}
\title{Run the AMCA algorithm.}
\usage{
amca(DATA, ResultsFolder, parameters, ObsIndicesNames = c("LAGTIME", "MAE",
  "NSHF", "NSLF"), selectedModels = NULL, warmup = NULL, verbose = TRUE,
  deltim = 1)
}
\arguments{
\item{DATA}{This is a data.frame containing the observed time series (zoo objects). It is structured into three columns: P containing precipitation, E containing potential evapotranspiration and Q containing streamflow discharge.}

\item{ResultsFolder}{Path to the folder containing results from MC simulations.}

\item{parameters}{is a data.frame containing all the parameter sets. This can be generated by \code{fuse::generateParameters()}.}

\item{ObsIndicesNames}{Names of model performance indices.}

\item{selectedModels}{(OPTIONAL) This is a table that contains at list 1 column named \code{mid} (list of model structures). Other informations can be added as additional columns but will be ignored (default = NULL).}

\item{warmup}{Percentage of initial time steps to ignore (default is 0.5 that corresponds to 50 percent of the full length).}

\item{verbose}{if set to TRUE it prints running information (default is FALSE).}

\item{deltim}{time step for FUSE simulations}
}
\value{
A list of suggested model configurations.
}
\description{
Run the AMCA algorithm.
}
\examples{
# results <- amca(DATA, ResultsFolder)

}

