% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amca.R
\name{amca}
\alias{amca}
\title{Run the AMCA algorithm.}
\usage{
amca(DATA, ResultsFolder, selectedModels = NULL, warmup = NULL,
  verbose = TRUE, PreSel = TRUE)
}
\arguments{
\item{DATA}{This is a data.frame containing the observed time series (zoo objects). It is structured into three columns: P containing precipitation, E containing potential evapotranspiration and Q containing streamflow discharge.}

\item{ResultsFolder}{Path to the folder containing results from MC simulations.}

\item{selectedModels}{(OPTIONAL) This is a table that contains at list 1 column named \code{mid} (list of model structures). Other informations can be added as additional columns but will be ignored (default = NULL).}

\item{warmup}{Percentage of initial time steps to ignore (default is 0.5 that corresponds to 50 percent of the full length).}

\item{verbose}{if set to TRUE it prints running information (default is FALSE).}

\item{PreSel}{if set to FALSE the preliminary selection step is skipped (default is TRUE).}
}
\value{
A list of suggested model configurations.
}
\description{
Run the AMCA algorithm.
}
\examples{
# results <- amca(DATA, ResultsFolder)

}

