% Generated by roxygen2 (4.0.2): do not edit by hand
\name{amca}
\alias{amca}
\title{Run the AMCA algorithm.}
\usage{
amca(DATA, parameters, MPIs, ResultsFolder, deltim, selectedModels = NULL,
  warmup = 0, verbose = FALSE, PreSel = TRUE, allBounds = FALSE)
}
\arguments{
\item{DATA}{This is a data.frame containing the observed time series (zoo objects).
It is structured into three columns:
"P" containing precipitation,
"E" containing evapo-transpiration and
"Q" containing streamflow discharge.}

\item{parameters}{This is a named data frame containing the parameter table, where each column corresponds to a parameter and each row to a realization.}

\item{MPIs}{list of functions describing the Model performance Indices.}

\item{ResultsFolder}{Path to the folder containing results from MC simulations.}

\item{deltim}{time step in days}

\item{selectedModels}{(OPTIONAL) This is a table that contains at list 1 column named "mid" (list of model structures). Other informations can be added as additional columns but will be ignored (default = NULL).}

\item{warmup}{Number of initial time steps to ignore (default = 0).}

\item{verbose}{if set to TRUE it prints running information (default = FALSE).}

\item{PreSel}{if set to FALSE the preliminary selection step is skipped (default = TRUE).}

\item{allBounds}{if set to TRUE it calculates ensembles of intermediate steps, if set to FALSE only BoundIE and BoundsRE are calculated (default = FALSE).}
}
\value{
A list of objects to infer.
}
\description{
Run the AMCA algorithm.
}
\examples{
# results <- damach( DATA, parameters, MPIs, ResultsFolder )
}

