% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BarPlots}
\alias{BarPlots}
\title{Create bar plots}
\usage{
BarPlots(mTable, selectedOption)
}
\arguments{
\item{mTable}{molt table}

\item{selectedOption}{a selection of options}
}
\value{
bar plot
}
\description{
Create bar plots
}
\examples{
# BarPlots(mTable,selectedOption)
}

