% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildEnsemble.R
\name{BuildEnsemble}
\alias{BuildEnsemble}
\title{Extracts the ensemble bounds from a given a set of realisations.}
\usage{
BuildEnsemble(observedQ, SimulationFolder, MIDs = NULL, PIDs = NULL,
  lowerP = 0.05, upperP = 0.95, verbose = FALSE, outputQ = FALSE,
  bigfile = FALSE)
}
\arguments{
\item{observedQ}{this is the observed streamflows}

\item{SimulationFolder}{path to the folder where simulations are stored}

\item{MIDs}{list of model structures to use}

\item{PIDs}{list of parameter indices to use}

\item{lowerP}{lower probability (e.g. 0.05 means 5th percentile)}

\item{upperP}{upper probability (e.g. 0.95 means 95th percentile)}

\item{verbose}{if set to TRUE it prints running information}

\item{outputQ}{if set to TRUE, the funtion also returns the discharges matrix}

\item{bigfile}{if set to TRUE, the funtion uses the bigmemory package}
}
\value{
A data.frame with 6 columns: date&time (Dates), observed discharge (Qobs), lower bound (lQ), median (mQ), upper bound (uQ), minimum (min) and maximum (max).
}
\description{
Results are generally divided in multiple files, therefore the function requires the simulations' folder path.
}
\examples{
# BuildEnsemble(observedQ, SimulationFolder, MIDs, PIDs)

}

