% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ParetoFrontier}
\alias{ParetoFrontier}
\title{Generate a table containing only non-dominated realisations based on MPIs and their Pareto Frontier.}
\usage{
ParetoFrontier(Indices, realisations, ObsIndices)
}
\arguments{
\item{Indices}{3d array containing indices (1000 parameter sets x 10 perf. measures x 252 model structures)}

\item{realisations}{list of model structures to take into account}

\item{ObsIndices}{this is the list of model performance indices calculated from observations.}
}
\value{
a subset of the initial dataset containing only non-dominated realisations
}
\description{
Generate a table containing only non-dominated realisations based on MPIs and their Pareto Frontier.
}
\examples{
# ParetoFrontier(Indices, PreSelTable, ObsIndices)
}

