% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotModelSimilarities.R
\name{PlotModelSimilarities}
\alias{PlotModelSimilarities}
\title{Plot model similarities (for comparison).}
\usage{
PlotModelSimilarities(ModelList, EnsembleTable, synMID = "",
  plotType = "frequency", myTitle = "", modellingOptions = NULL)
}
\arguments{
\item{ModelList}{This is a table which contains at least 1 colum named "mid". Other informations can be added as additional columns but will be ignored.}

\item{EnsembleTable}{table containing alle the combination of mid and pid selected so far}

\item{synMID}{if running synthetic experiments, this is the synthetic MID number (otherwise leave it empty)}

\item{plotType}{default is "frequency" (alternative: "persistence")}

\item{myTitle}{title of the plot}

\item{modellingOptions}{is a vector of labels to specify which modelling options should be used. If set to NULL, it uses all the modelling options.}
}
\value{
A plot comparing model similarities.
}
\description{
Plot model similarities (for comparison).
}
\examples{
# PlotModelSimilarities(ModelList,results$RETable)
}

