% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotParameterSimilarities.R
\name{PlotParameterSimilarities}
\alias{PlotParameterSimilarities}
\title{Plot prior and posterior parameter distributions (for comparison).}
\usage{
PlotParameterSimilarities(RealisationsTable, parameters,
  selectedParams = NULL, synparameters = NULL, labelPRE = "PRIOR",
  labelPOST = "POSTERIOR")
}
\arguments{
\item{RealisationsTable}{list of the filtered realisations}

\item{parameters}{This is a named data frame containing the parameter table, where each column corresponds to a parameter and each row to a realization.}

\item{selectedParams}{is a list of pid to filter}

\item{synparameters}{is the synthetic parameter set (if available)}

\item{labelPRE}{label for the prior}

\item{labelPOST}{label for the posterior}
}
\value{
A plot comparing prior and posterior parameter distributions.
}
\description{
Plot prior and posterior parameter distributions (for comparison).
}
\examples{
# PlotParameterSimilarities(RealisationsTable, parameters)

}

