% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RedundancyReduction}
\alias{RedundancyReduction}
\title{Redundancy reduction process using Dynamic Time Warping.}
\usage{
RedundancyReduction(ParetoFront, DATA, the.som, ParameterSet, observedQ, deltim,
  pperiod, verbose = FALSE)
}
\arguments{
\item{ParetoFront}{Pareto Frontier}

\item{DATA}{this is the data frame containing the observations}

\item{the.som}{som map}

\item{ParameterSet}{is the table of parameters}

\item{observedQ}{observed discharge (warmup removed)}

\item{deltim}{time step in days}

\item{pperiod}{is the vector of time step to take into consideration when calculating performances}

\item{verbose}{if set to TRUE it prints running information (default = FALSE)}
}
\value{
A subset of the table containing the filtered configurations.
}
\description{
Redundancy reduction process using Dynamic Time Warping.
}
\details{
This function removes redundant realizations by extracting the realization with the best DTW score from each cluster.
}
\examples{
# RedundancyReduction(ParetoFrontTable,DATA,the.som,parameters,observedQ,deltim,pperiod)
}

