% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RescaleIndices.R
\name{RescaleIndices}
\alias{RescaleIndices}
\title{Rescale Model Performance Indices (MPIs) between 0 and 1.}
\usage{
RescaleIndices(Indices, ObsIndices)
}
\arguments{
\item{Indices}{this is the array containing the not-yet rescaled Model Performance Indices.}

\item{ObsIndices}{this is the list of model performance indices calculated from observations.}
}
\value{
Rescaled array of Model Performance Indices
}
\description{
Rescale Model Performance Indices (MPIs) between 0 and 1.
}
\details{
The best performance always tends to zero.
}
\examples{
# newIndices <- RescaleIndices(Indices, ObsIndices)
}

