% Generated by roxygen2 (4.0.2): do not edit by hand
\name{review}
\alias{review}
\title{Summarise performance coefficients}
\usage{
review(IE, RE, observedQ, type = "B")
}
\arguments{
\item{IE}{Initial Ensemble bounds}

\item{RE}{Reduced Ensemble bounds and discharges}

\item{observedQ}{observed discharge (warmup removed)}

\item{type}{if equal to "B", only the max/min bounds are used. If equal to "P", it uses the upper/lower percentiles.}
}
\value{
Returns accuracy, precision and statistical reliability
}
\description{
Summarise performance coefficients
}
\examples{
# review(IE, RE)
}

